Ура деление и тут работает

+++ task1/2.3/calc.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 
 
-int add(int a, int b) {
-    return a + b;
+double add(double a, double b) {
+    return (a + b);
 }
 
-int subtract(int a, int b) {
-    return a - b;
+double subtract(double a, double b) {
+    return (a - b);
 }
 
-int multiply(int a, int b) {
-    return a * b;
+double multiply(double a, double b) {
+    return (a * b);
 }
 
-double divide(int a, int b) {
+double divide(double a, double b) {
     if (b == 0) {
         printf("Ошибка: Нельзя делить на ноль.\n");
         return 0.0;
     }
-    return (double)a / (double)b;
+    return ((double)a / (double)b);
 }
